server/requirements.txt
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.7
cheroot==8.6.0
CherryPy==18.6.1
click==8.0.3
cycler==0.10.0
dnspython==2.3.0
et-xmlfile==1.1.0
Flask==2.0.2
idna==3.3
itsdangerous==2.0.1
jaraco.classes==3.2.1
jaraco.collections==3.5.1
jaraco.context==4.1.1
jaraco.functools==3.5.0
jaraco.text==3.8.0
Jinja2==3.0.2
kiwisolver==1.3.2
MarkupSafe==2.0.1
numpy==1.21.3
openai==0.19.0
openpyxl==3.0.9
pandas==1.3.4
pandas-stubs==1.2.0.35
Pillow==8.3.2
portend==3.1.0
protobuf==3.17.3
pycodestyle==2.8.0
pycrypto==2.6.1
pymongo==4.3.3
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==5.4.1
repoze.lru==0.7
requests==2.26.0
Routes==2.5.1
six==1.16.0
tempora==5.0.1
toml==0.10.2
tqdm==4.62.3
urllib3==1.26.7
Werkzeug==2.0.2
zc.lockfile==2.0
